pyproject.toml
[project]
name = "dna-synthesizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "biopython>=1.86",
    "einops>=0.8.2",
    "hydra-core>=1.3.2",
    "torch>=2.11.0",
]
